refactor: rewrite toggle stories
