+++ src/components/Toggle/__stories__/Disabled.stories.tsx
@@ -1,0 +1,9 @@
+import { Template } from './Template.stories'
+
+export const Disabled = Template.bind({})
+
+Disabled.args = {
+  label: 'Toggle me on',
+  name: 'label',
+  disabled: true,
+}
