+++ src/components/Toggle/__stories__/Label.stories.tsx
@@ -1,0 +1,5 @@
+import { Template } from './Template.stories'
+
+export const Label = Template.bind({})
+
+Label.args = { label: 'Toggle me on', name: 'label' }
