+++ src/components/Toggle/__stories__/Playground.stories.tsx
@@ -1,0 +1,5 @@
+import { Template } from './Template.stories'
+
+export const Playground = Template.bind({})
+
+Playground.args = { ...Template.args }
