+++ src/components/Toggle/__stories__/Template.stories.tsx
@@ -1,0 +1,6 @@
+import { ComponentStory } from '@storybook/react'
+import Toggle from '..'
+
+export const Template: ComponentStory<typeof Toggle> = args => (
+  <Toggle {...args} />
+)
