+++ src/components/Toggle/__stories__/Tooltip.stories.tsx
@@ -1,0 +1,9 @@
+import { Template } from './Template.stories'
+
+export const Tooltip = Template.bind({})
+
+Tooltip.args = {
+  label: 'Toggle me on',
+  name: 'label',
+  tooltip: 'Hello there!',
+}
